nit: don't need to keep `tabId` reference across try/catch
The first tabId is set only if step above it succeeds. If it's not
reached, we'd need new tab anyway. If not, completeGrant will
redirectToWelcomeScreen for that tab only. So, that tab isn't important
to us in "catch" block anymore.
i.e. in the catch, it'll be never already defined.